fix: use a saner way to parallelize

+++ threaded_apply.h
@@ -1,5 +1,5 @@
 #ifndef APPLYTHREADED_H
 #define APPLYTHREADED_H
 #include "def.h"
-int apply_threaded(BinaryFunction function, int *input, size_t size);
+int apply_threaded(BinaryFunction function, int *input, int size);
 #endif
